Restructure files and directories of store
